requirements.txt
# Video Topic Explainer — Python dependencies
# Tested on Python 3.10 / 3.11 / 3.12
#
# IMPORTANT: Do NOT install tensorflow or keras — they conflict with transformers.
# Install in this order:
#   1. torch (CPU)
#   2. transformers + sentence-transformers
#   3. everything else
#
# Quick install:
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
#   pip install -r requirements.txt

# ── Deep Learning (CPU-only torch, no tensorflow) ─────────────────────────────
torch>=2.1.0                   # CPU build — install separately (see above)
transformers==4.40.2           # Must be 4.40.x — newer versions need keras3
sentence-transformers==2.7.0   # Must be 2.7.x to match transformers==4.40.2

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Streamlit Frontend ────────────────────────────────────────────────────────
streamlit==1.35.0

# ── Audio Processing ──────────────────────────────────────────────────────────
ffmpeg-python==0.2.0

# ── Speech-to-Text ────────────────────────────────────────────────────────────
faster-whisper==1.0.1

# ── Topic Extraction ──────────────────────────────────────────────────────────
keybert==0.8.4

# ── LLM Provider ─────────────────────────────────────────────────────────────
groq==0.9.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.7.1